Rename AutoBoot.py to process.py
