+++ sardine/superdirt/__init__.py
@@ -1,1 +1,1 @@
-from .AutoBoot import *
+from .process import *
